chore: split home page stats into two sections & combine both stat sections & transaction table into one file
Intention is to display the smaller less important stats below the existing stat-transaction grid section. Both stat sections requires similar logic so I included the existing stat section with the new one together with the transaction table (as they are in a grid together).
